report: minor touchups and fix gatk/sention pr_curve bug
